python_LLM/pyproject.toml
[project]
name = "python-llm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "streamlit>=1.45.1",
    "openai",
    "python-dotenv"
]
